LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-img-element/update-the-image-data/current-request-microtask-expected.txt


FAIL currentSrc is updated only after the microtask that updates the current request is run assert_equals: after assigning to img.src but before the corresponding microtask is run expected "" but got "data:,c"
